+++ test/cuda/cuda.jl
@@ -13,7 +13,7 @@ using LinearAlgebra: I, cholesky, Cholesky
 
   x = Flux.onehotbatch([1, 2, 3], 1:3)
   cx = gpu(x)
-  @test cx isa Flux.OneHotMatrix && cx.data isa CuArray
+  @test cx isa Flux.OneHotMatrix && cx.indices isa CuArray
   @test (cx .+ 1) isa CuArray
 
   m = Chain(Dense(10, 5, tanh), Dense(5, 2), softmax)
@@ -38,10 +38,23 @@ using LinearAlgebra: I, cholesky, Cholesky
 
 end
 
+@testset "onehot gpu" begin
+  y = Flux.onehotbatch(ones(3), 1:2) |> gpu;
+  @test (repr("text/plain", y); true)
+end
+
 @testset "onecold gpu" begin
   y = Flux.onehotbatch(ones(3), 1:10) |> gpu;
+  l = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j']
   @test Flux.onecold(y) isa CuArray
   @test y[3,:] isa CuArray
+  @test Flux.onecold(y, l) == ['a', 'a', 'a']
+end
+
+@testset "onehot forward map to broadcast" begin
+  oa = OneHotArray(rand(1:10, 5, 5), 10) |> gpu
+  @test all(map(identity, oa) .== oa)
+  @test all(map(x -> 2 * x, oa) .== 2 .* oa)
 end
 
 @testset "restructure gpu" begin
@@ -59,5 +72,15 @@ end
     @test Q_gpu isa Cholesky{<:Any,<:CuArray}
     Q_cpu = Q_gpu |> cpu
     @test Q_cpu == cholesky(eltype(Q_gpu).(M))
+  end
+
+  @testset "isbits array types" begin
+    struct SimpleBits
+      field::Int32
+    end
+    
+    @test gpu((;a=ones(1))).a isa CuVector{Float32}
+    @test gpu((;a=['a', 'b', 'c'])).a isa CuVector{Char}
+    @test gpu((;a=[SimpleBits(1)])).a isa CuVector{SimpleBits}
   end
 end